test: split Windows-qualified-path release-topology tests into a sibling module
Codex review on PR #1175: the Windows $PWD fix grew
tests/test_action_release_topology_config.py from 1,181 to 1,213 lines,
past the architecture gate's 1,200-line test-module cap
([new-test-size] on scripts/check_architecture.py).

Moved TestReleaseTopologyOverlayPreservesWindowsQualifiedBuildConfigPath
(and its two dedicated helpers, _base_source_absolutize_source and
_bash_pwd) into a new tests/test_action_release_topology_windows_paths.py,
duplicating the small set of shared bash-harness primitives
(RUN_SH, _path_qualified_helper_source, _bash_executable,
_run_bash_script) rather than importing them from the parent module --
mirrors this directory's established convention of not cross-importing
test code between bash-harness test files (e.g.
test_action_compile_context_parity.py duplicates its own
_bash_executable/_run_bash_script rather than importing them).

Both files are now well under the cap (1074 and 245 lines). All 34
tests across the two files still pass.
